# docs/swagger.yaml
basePath: /api
definitions:
...
  chat.SendMessageRequest:
    properties:
      message_text:
        example: Привет! Интересует ваше предложение.
        type: string
      message_type:
        example: text
        type: string
    required:
    - message_text
    type: object
  chat.SendMessageResponse:
    properties:
      message:
        $ref: '#/definitions/models.ChatMessageResponse'
      success:
        type: boolean
    type: object
...
  models.ChatMessageResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_from_me:
        type: boolean
      is_read:
        type: boolean
      message_text:
        type: string
      message_type:
        type: string
      read_at:
        type: string
      sender_id:
        type: integer
      sender_name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API для приложения MoveShare
  title: MoveShare API
  version: "1.0"
paths:
...
  /chats/{chatId}/mark-read:
    post:
      description: Mark all unread messages in a chat as read for the current user
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark messages as read
      tags:
      - Chat
  /chats/{chatId}/messages:
    get:
      description: Returns messages for a specific chat with pagination
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - default: 30
        description: Limit number of messages returned
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: desc
        description: Order of messages
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get chat messages
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: Sends a new message to the specified chat
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/chat.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chat.SendMessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send message to chat
      tags:
      - Chat
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in format: Bearer {your_token_here}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"